Safe interrupt processing and Eliminate String class usage
*  Rework I2C Fona message handling.  The receiveEvent just copies the
message into a fonaMsg array and sets a flag.  This moves most of the
processing out of the interrupt routine.  It also avoids corruption due
to String and overlapping messages.

*  Compiling a response now uses the same print methods as Serial, but
they are "printed" to the reponse RAM array.  (requestEvent is prevented
during this print process.)  Then those characters are sent.